feat: use `f32` instead of `f64`

+++ src/engine/constants.rs
@@ -1,11 +1,11 @@
-pub const DEFAULT_COLLISION_DAMPING_FACTOR: f64 = 6.0 / 10.0;
+pub const DEFAULT_COLLISION_DAMPING_FACTOR: f32 = 6.0 / 10.0;
 
-pub const GROUND_DRAG_FACTOR: f64 = 0.06;
+pub const GROUND_DRAG_FACTOR: f32 = 0.06;
 
-pub const AIR_RESISTANCE_FACTOR: f64 = 0.1;
+pub const AIR_RESISTANCE_FACTOR: f32 = 0.1;
 
-pub const GRAVITY: f64 = 100.0;
+pub const GRAVITY: f32 = 100.0;
 
 pub const FPS: i8 = 120;
 
-pub const DT: f64 = 1.0 / FPS as f64;
+pub const DT: f32 = 1.0 / FPS as f32;
